Added logic to Tooltips component & moved to demo-components
