+++ projects/demo-components/src/index.jsx
@@ -6,5 +6,5 @@ export { default as Header } from './header'
 export { default as Map } from './map'
 export { default as accessToken } from './access-token'
 export { default as PageShell } from './page-shell'
-export { default as MapboxTooltip } from './mapbox-tooltip'
+export { default as MapboxTooltips } from './mapbox-tooltips'
 export { default as LogoSVG } from './logo-svg'
